Feature: Big change!
Added message 'checking' to the system to use another LLM instance ensuring the output of the chat LLM is 'OK'.
Fixed the stream category change event, as well as renaming everything to 'category' rather than 'game' to remove gaming-centric language.
Updated documentation to match, and removed the entirety of the development documentation. The code changes too quickly for this to be useful right now.
Removed ALL test for the same reason. These will all be added at a later date.

A lot of removals, a lot of cleanup, some new additions.